llm-agents-lib/pyproject.toml
[project]
name = "nursing-homes-agent-lib"
version = "0.1.0"
description = "Library of agents for the nursing homes data text to SQL project"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.2.2",
    "duckdb-engine",
    "grandalf>=0.8",
    "langchain-community>=0.3.22",
    "langchain[google-genai]>=0.3.24",
    "langchainhub>=0.1.21",
    "langgraph>=0.3.34",
    "plotly>=6.1.0",
    "polars>=1.29.0",
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
]

[tool.uv.sources]
# had to use local package, to fix bug with db name not extracted from path with multiple "/" chars
duckdb-engine = { path = "../../duckdb_engine", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "hatchling>=1.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/nursing_homes_agent_lib"]
